Fix rebase-onto-master conflict resolution mistakes and stale imports left by class relocations
