+++ its/pom.xml
@@ -15,7 +15,6 @@
   <description>Code Analyzer for Java :: Integration Tests :: Parent POM</description>
 
   <modules>
-    <module>autoscan</module>
     <module>scanner-integration-tests</module>
     <module>plugin</module>
     <module>ruling</module>
