+++ its/scanner-integration-tests/pom.xml
@@ -5,7 +5,7 @@
   <parent>
     <groupId>org.sonarsource.java</groupId>
     <artifactId>java-its</artifactId>
-    <version>8.32.0-SNAPSHOT</version>
+    <version>8.42.0-SNAPSHOT</version>
   </parent>
 
   <artifactId>it-java-scanner-integration-tests</artifactId>
