+++ java-frontend/src/test/java/org/sonar/java/model/springcontext/SpringContextGathererTest.java
@@ -16,9 +16,9 @@
  */
 package org.sonar.java.model.springcontext;
 
+import com.sonarsource.scanner.engine.sensor.test.fixtures.SensorContextTester;
 import java.io.File;
 import java.util.List;
-import org.sonar.api.batch.sensor.internal.SensorContextTester;
 import org.sonar.java.SonarComponents;
 import org.sonar.java.TestUtils;
 import org.sonar.java.model.JParserTestUtils;
